+++ internal/testdata/log.go
@@ -44,7 +44,7 @@ func GenerateLogs(count int) plog.Logs {
 func fillLogOne(log plog.LogRecord) {
 	log.SetTimestamp(logTimestamp)
 	log.SetDroppedAttributesCount(1)
-	log.SetSeverityNumber(plog.SeverityNumberINFO)
+	log.SetSeverityNumber(plog.SeverityNumberInfo)
 	log.SetSeverityText("Info")
 	log.SetSpanID(pcommon.NewSpanID([8]byte{0x01, 0x02, 0x04, 0x08}))
 	log.SetTraceID(pcommon.NewTraceID([16]byte{0x08, 0x04, 0x02, 0x01}))
@@ -59,7 +59,7 @@ func fillLogOne(log plog.LogRecord) {
 func fillLogTwo(log plog.LogRecord) {
 	log.SetTimestamp(logTimestamp)
 	log.SetDroppedAttributesCount(1)
-	log.SetSeverityNumber(plog.SeverityNumberINFO)
+	log.SetSeverityNumber(plog.SeverityNumberInfo)
 	log.SetSeverityText("Info")
 
 	attrs := log.Attributes()