+++ pdata/ptrace/json.go
@@ -16,15 +16,14 @@ package ptrace // import "go.opentelemetry.io/collector/pdata/ptrace"
 
 import (
 	"bytes"
-	"encoding/base64"
 	"fmt"
 
 	"github.com/gogo/protobuf/jsonpb"
 	jsoniter "github.com/json-iterator/go"
 
 	"go.opentelemetry.io/collector/pdata/internal"
-	otlpcommon "go.opentelemetry.io/collector/pdata/internal/data/protogen/common/v1"
 	otlptrace "go.opentelemetry.io/collector/pdata/internal/data/protogen/trace/v1"
+	"go.opentelemetry.io/collector/pdata/internal/json"
 )
 
 // NewJSONMarshaler returns a model.Marshaler. Marshals to OTLP json bytes.
@@ -90,11 +89,11 @@ func readResourceSpans(iter *jsoniter.Iterator) *otlptrace.ResourceSpans {
 				switch f {
 				case "attributes":
 					iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
-						rs.Resource.Attributes = append(rs.Resource.Attributes, readAttribute(iter))
+						rs.Resource.Attributes = append(rs.Resource.Attributes, json.ReadAttribute(iter))
 						return true
 					})
 				case "droppedAttributesCount", "dropped_attributes_count":
-					rs.Resource.DroppedAttributesCount = iter.ReadUint32()
+					rs.Resource.DroppedAttributesCount = json.ReadUint32(iter)
 				default:
 					iter.ReportError("readResourceSpans.resource", fmt.Sprintf("unknown field:%v", f))
 				}
@@ -172,30 +171,30 @@ func readSpan(iter *jsoniter.Iterator) *otlptrace.Span {
 		case "kind":
 			sp.Kind = readSpanKind(iter)
 		case "startTimeUnixNano", "start_time_unix_nano":
-			sp.StartTimeUnixNano = uint64(readInt64(iter))
+			sp.StartTimeUnixNano = json.ReadUint64(iter)
 		case "endTimeUnixNano", "end_time_unix_nano":
-			sp.EndTimeUnixNano = uint64(readInt64(iter))
+			sp.EndTimeUnixNano = json.ReadUint64(iter)
 		case "attributes":
 			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
-				sp.Attributes = append(sp.Attributes, readAttribute(iter))
+				sp.Attributes = append(sp.Attributes, json.ReadAttribute(iter))
 				return true
 			})
 		case "droppedAttributesCount", "dropped_attributes_count":
-			sp.DroppedAttributesCount = iter.ReadUint32()
+			sp.DroppedAttributesCount = json.ReadUint32(iter)
 		case "events":
 			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
 				sp.Events = append(sp.Events, readSpanEvent(iter))
 				return true
 			})
 		case "droppedEventsCount", "dropped_events_count":
-			sp.DroppedEventsCount = iter.ReadUint32()
+			sp.DroppedEventsCount = json.ReadUint32(iter)
 		case "links":
 			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
 				sp.Links = append(sp.Links, readSpanLink(iter))
 				return true
 			})
 		case "droppedLinksCount", "dropped_links_count":
-			sp.DroppedLinksCount = iter.ReadUint32()
+			sp.DroppedLinksCount = json.ReadUint32(iter)
 		case "status":
 			iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
 				switch f {
@@ -233,11 +232,11 @@ func readSpanLink(iter *jsoniter.Iterator) *otlptrace.Span_Link {
 			link.TraceState = iter.ReadString()
 		case "attributes":
 			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
-				link.Attributes = append(link.Attributes, readAttribute(iter))
+				link.Attributes = append(link.Attributes, json.ReadAttribute(iter))
 				return true
 			})
 		case "droppedAttributesCount", "dropped_attributes_count":
-			link.DroppedAttributesCount = iter.ReadUint32()
+			link.DroppedAttributesCount = json.ReadUint32(iter)
 		default:
 			iter.ReportError("readSpanLink", fmt.Sprintf("unknown field:%v", f))
 		}
@@ -252,137 +251,22 @@ func readSpanEvent(iter *jsoniter.Iterator) *otlptrace.Span_Event {
 	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
 		switch f {
 		case "timeUnixNano", "time_unix_nano":
-			event.TimeUnixNano = uint64(readInt64(iter))
+			event.TimeUnixNano = json.ReadUint64(iter)
 		case "name":
 			event.Name = iter.ReadString()
 		case "attributes":
 			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
-				event.Attributes = append(event.Attributes, readAttribute(iter))
+				event.Attributes = append(event.Attributes, json.ReadAttribute(iter))
 				return true
 			})
 		case "droppedAttributesCount", "dropped_attributes_count":
-			event.DroppedAttributesCount = iter.ReadUint32()
+			event.DroppedAttributesCount = json.ReadUint32(iter)
 		default:
 			iter.ReportError("readSpanEvent", fmt.Sprintf("unknown field:%v", f))
 		}
 		return true
 	})
 	return event
-}
-
-func readAttribute(iter *jsoniter.Iterator) otlpcommon.KeyValue {
-	kv := otlpcommon.KeyValue{}
-	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
-		switch f {
-		case "key":
-			kv.Key = iter.ReadString()
-		case "value":
-			iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
-				kv.Value = readAnyValue(iter, f)
-				return true
-			})
-		default:
-			iter.ReportError("readAttribute", fmt.Sprintf("unknown field:%v", f))
-		}
-		return true
-	})
-	return kv
-}
-
-func readAnyValue(iter *jsoniter.Iterator, f string) otlpcommon.AnyValue {
-	switch f {
-	case "stringValue", "string_value":
-		return otlpcommon.AnyValue{
-			Value: &otlpcommon.AnyValue_StringValue{
-				StringValue: iter.ReadString(),
-			},
-		}
-	case "boolValue", "bool_value":
-		return otlpcommon.AnyValue{
-			Value: &otlpcommon.AnyValue_BoolValue{
-				BoolValue: iter.ReadBool(),
-			},
-		}
-	case "intValue", "int_value":
-		return otlpcommon.AnyValue{
-			Value: &otlpcommon.AnyValue_IntValue{
-				IntValue: readInt64(iter),
-			},
-		}
-	case "doubleValue", "double_value":
-		return otlpcommon.AnyValue{
-			Value: &otlpcommon.AnyValue_DoubleValue{
-				DoubleValue: iter.ReadFloat64(),
-			},
-		}
-	case "bytesValue", "bytes_value":
-		v, err := base64.StdEncoding.DecodeString(iter.ReadString())
-		if err != nil {
-			iter.ReportError("bytesValue", fmt.Sprintf("base64 decode:%v", err))
-			return otlpcommon.AnyValue{}
-		}
-		return otlpcommon.AnyValue{
-			Value: &otlpcommon.AnyValue_BytesValue{
-				BytesValue: v,
-			},
-		}
-	case "arrayValue", "array_value":
-		return otlpcommon.AnyValue{
-			Value: &otlpcommon.AnyValue_ArrayValue{
-				ArrayValue: readArray(iter),
-			},
-		}
-	case "kvlistValue", "kvlist_value":
-		return otlpcommon.AnyValue{
-			Value: &otlpcommon.AnyValue_KvlistValue{
-				KvlistValue: readKvlistValue(iter),
-			},
-		}
-	default:
-		iter.ReportError("readAnyValue", fmt.Sprintf("unknown field:%v", f))
-		return otlpcommon.AnyValue{}
-	}
-}
-
-func readArray(iter *jsoniter.Iterator) *otlpcommon.ArrayValue {
-	v := &otlpcommon.ArrayValue{}
-	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
-		switch f {
-		case "values":
-			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
-				iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
-					v.Values = append(v.Values, readAnyValue(iter, f))
-					return true
-				})
-				return true
-			})
-		default:
-			iter.ReportError("readArray", fmt.Sprintf("unknown field:%s", f))
-		}
-		return true
-	})
-	return v
-}
-
-func readKvlistValue(iter *jsoniter.Iterator) *otlpcommon.KeyValueList {
-	v := &otlpcommon.KeyValueList{}
-	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
-		switch f {
-		case "values":
-			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
-				v.Values = append(v.Values, readAttribute(iter))
-				return true
-			})
-		default:
-			iter.ReportError("readKvlistValue", fmt.Sprintf("unknown field:%s", f))
-		}
-		return true
-	})
-	return v
-}
-
-func readInt64(iter *jsoniter.Iterator) int64 {
-	return iter.ReadAny().ToInt64()
 }
 
 func readSpanKind(iter *jsoniter.Iterator) otlptrace.Span_SpanKind {