+++ processor/processorhelper/traces.go
@@ -36,18 +36,20 @@ type tracesProcessor struct {
 	consumer.Traces
 }
 
-// Deprecated: [v0.58.0] use version with NewTracesProcessorWithCreateSettings.
-func NewTracesProcessor(
+// Deprecated: [v0.58.0] use version with NewTracesProcessor.
+func NewTracesProcessorWithCreateSettings(
+	ctx context.Context,
+	set component.ProcessorCreateSettings,
 	cfg config.Processor,
 	nextConsumer consumer.Traces,
 	tracesFunc ProcessTracesFunc,
 	options ...Option,
 ) (component.TracesProcessor, error) {
-	return NewTracesProcessorWithCreateSettings(context.Background(), component.ProcessorCreateSettings{}, cfg, nextConsumer, tracesFunc, options...)
+	return NewTracesProcessor(ctx, set, cfg, nextConsumer, tracesFunc, options...)
 }
 
-// NewTracesProcessorWithCreateSettings creates a TracesProcessor that ensure context propagation and the right tags are set.
-func NewTracesProcessorWithCreateSettings(
+// NewTracesProcessor creates a component.TracesProcessor that ensure context propagation and the right tags are set.
+func NewTracesProcessor(
 	_ context.Context,
 	_ component.ProcessorCreateSettings,
 	cfg config.Processor,