+++ service/collector_test.go
@@ -105,6 +105,49 @@ func TestCollectorCancelContext(t *testing.T) {
 	assert.Equal(t, Closed, col.GetState())
 }
 
+type mockCfgProvider struct {
+	ConfigProvider
+	watcher chan error
+}
+
+func (p mockCfgProvider) Watch() <-chan error {
+	return p.watcher
+}
+
+func TestCollectorStateAfterConfigChange(t *testing.T) {
+	factories, err := componenttest.NopFactories()
+	require.NoError(t, err)
+
+	provider, err := NewConfigProvider(newDefaultConfigProviderSettings([]string{filepath.Join("testdata", "otelcol-nop.yaml")}))
+	require.NoError(t, err)
+
+	watcher := make(chan error, 1)
+	col, err := New(CollectorSettings{
+		BuildInfo:      component.NewDefaultBuildInfo(),
+		Factories:      factories,
+		ConfigProvider: &mockCfgProvider{ConfigProvider: provider, watcher: watcher},
+		telemetry:      newColTelemetry(featuregate.NewRegistry()),
+	})
+	require.NoError(t, err)
+
+	wg := startCollector(context.Background(), t, col)
+
+	assert.Eventually(t, func() bool {
+		return Running == col.GetState()
+	}, 2*time.Second, 200*time.Millisecond)
+
+	watcher <- nil
+
+	assert.Eventually(t, func() bool {
+		return Running == col.GetState()
+	}, 2*time.Second, 200*time.Millisecond)
+
+	col.Shutdown()
+
+	wg.Wait()
+	assert.Equal(t, Closed, col.GetState())
+}
+
 func TestCollectorReportError(t *testing.T) {
 	factories, err := componenttest.NopFactories()
 	require.NoError(t, err)
@@ -127,6 +170,33 @@ func TestCollectorReportError(t *testing.T) {
 	}, 2*time.Second, 200*time.Millisecond)
 
 	col.service.host.ReportFatalError(errors.New("err2"))
+
+	wg.Wait()
+	assert.Equal(t, Closed, col.GetState())
+}
+
+func TestCollectorSendSignal(t *testing.T) {
+	factories, err := componenttest.NopFactories()
+	require.NoError(t, err)
+
+	cfgProvider, err := NewConfigProvider(newDefaultConfigProviderSettings([]string{filepath.Join("testdata", "otelcol-nop.yaml")}))
+	require.NoError(t, err)
+
+	col, err := New(CollectorSettings{
+		BuildInfo:      component.NewDefaultBuildInfo(),
+		Factories:      factories,
+		ConfigProvider: cfgProvider,
+		telemetry:      newColTelemetry(featuregate.NewRegistry()),
+	})
+	require.NoError(t, err)
+
+	wg := startCollector(context.Background(), t, col)
+
+	assert.Eventually(t, func() bool {
+		return Running == col.GetState()
+	}, 2*time.Second, 200*time.Millisecond)
+
+	col.signalsChannel <- syscall.SIGTERM
 
 	wg.Wait()
 	assert.Equal(t, Closed, col.GetState())
@@ -321,7 +391,8 @@ func testCollectorStartHelper(t *testing.T, telemetry *telemetryInitializer, tc 
 	assertMetrics(t, metricsAddr, tc.expectedLabels)
 
 	assertZPages(t)
-	col.signalsChannel <- syscall.SIGTERM
+
+	col.Shutdown()
 
 	wg.Wait()
 	assert.Equal(t, Closed, col.GetState())
@@ -341,6 +412,47 @@ func TestCollectorStartWithOpenTelemetryMetrics(t *testing.T) {
 			colTel := newColTelemetry(featuregate.NewRegistry())
 			require.NoError(t, colTel.registry.Apply(map[string]bool{useOtelForInternalMetricsfeatureGateID: true}))
 			testCollectorStartHelper(t, colTel, tc)
+		})
+	}
+}
+
+func TestCollectorStartWithTraceContextPropagation(t *testing.T) {
+	tests := []struct {
+		file        string
+		errExpected bool
+	}{
+		{file: "otelcol-invalidprop.yaml", errExpected: true},
+		{file: "otelcol-nop.yaml", errExpected: false},
+		{file: "otelcol-validprop.yaml", errExpected: false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.file, func(t *testing.T) {
+			factories, err := componenttest.NopFactories()
+			require.NoError(t, err)
+
+			cfgProvider, err := NewConfigProvider(newDefaultConfigProviderSettings([]string{filepath.Join("testdata", tt.file)}))
+			require.NoError(t, err)
+
+			set := CollectorSettings{
+				BuildInfo:      component.NewDefaultBuildInfo(),
+				Factories:      factories,
+				ConfigProvider: cfgProvider,
+				telemetry:      newColTelemetry(featuregate.NewRegistry()),
+			}
+
+			col, err := New(set)
+			require.NoError(t, err)
+
+			if tt.errExpected {
+				require.Error(t, col.Run(context.Background()))
+				assert.Equal(t, Closed, col.GetState())
+			} else {
+				wg := startCollector(context.Background(), t, col)
+				col.Shutdown()
+				wg.Wait()
+				assert.Equal(t, Closed, col.GetState())
+			}
 		})
 	}
 }